vaccinations: copy updates
